Harden the day boundary, the diary, and every degraded state
An audit of the whole surface — security, backend correctness, client
flows — turned up four ways the app could lose or corrupt a user's data.
All four are fixed here, with the regression tests that catch them.

The day could move backward. Flying Kolkata -> New York rewinds the local
date, and every write that day then landed on an already-finished day:
meals appended to yesterday, its totals re-summed and inflated, its
completions re-derived, and the real day orphaned where nothing could
reach it. The resolved day is now ratcheted — forward at midnight or
travelling east, never back.

Workout sets were filed by the DATABASE's day, not the user's. A set
logged at 18:00 in Los Angeles carried tomorrow's UTC date, so
/today-sets returned {} and the page restarted the session at zero. People
re-logged, and the volume double-counted. workout_logs now carries a
user-local date like every other date-keyed table.

The diary destroyed typed figures. "The diary total IS the day's figure"
held right up until someone typed 2200 kcal for a restaurant day and then
logged one 250 kcal snack — the day silently became 250, and the AI
narrated it from the wrong number. values_source records who owns each
figure; once you type one, the diary stops overwriting it that day.

Concurrent checklist edits lost each other. custom_items was a
read-modify-write on a jsonb array, so a second tab's write erased the
first behind a 200. The three mutations now happen inside the UPDATE.

Also: adherence reported 0% for a diligent user whose rows were dated
ahead of today; a day whose snapshot build failed could never self-heal,
leaving 1g of protein enough to complete it; a profile goal change didn't
reach today's mission; briefing.date came back UTC-shifted on some paths
and YYYY-MM-DD on others; null + 2.5 told people to add weight to their
push-ups.

Security: the plan schema had no length or array bounds, and the plan is
re-parsed on nearly every request — one oversized plan permanently taxed
that account and the shared pool. /health's unauthenticated DB ping is
memoized so it can't be used to starve the connection pool for free.

Client: a failed BACKGROUND refresh no longer wipes the Progress page —
a 429 on tab-focus made a user's history look deleted. The briefing card
no longer vanishes silently on error, Profile no longer dead-ends on a
404 it can't retry out of, and the diary no longer shows "nothing logged
yet" over a failed load, which invited people to re-log meals they
already had. Errors carry a status now instead of being matched by regex
against server prose — one reworded string could have pushed established
users back through signup and restarted their goal clock.

Docs were rebuilt against the actual route table: five documented
endpoints did not exist and seven real ones were missing. A contract test
now fails the build if they drift again.

Tests: 45 -> 56, plus six new end-to-end invariants. The expiringMap
suite had been silently excluded from `npm test` and now runs.

+++ fitai/server/src/validators/requestSchemas.js
@@ -63,7 +63,10 @@ const ProfileUpdateSchema = OnboardingSchema.partial().refine(
 const MealSchema = z.object({
   name: z.string().min(1).max(120),
   grams: z.number().positive().max(2000).optional(),
-  calories: z.number().int().min(0).max(5000),
+  // Not .int(): the vision path returns fractional kcal (FoodItemSchema
+  // allows them) and the manual input has no step, so "250.5" would 400 on
+  // a number the app itself can produce. Rounded on save instead.
+  calories: z.number().min(0).max(5000),
   protein: z.number().min(0).max(300).optional().default(0),
   carbs: z.number().min(0).max(500).optional(),
   fat: z.number().min(0).max(300).optional(),
